+++ radiant/points.cpp
@@ -178,8 +178,8 @@ void Pointfile_UpdateViews( CPointfile::const_iterator i ){
 	{
 		Vector3 dir( vector3_normalised( vector3_subtracted( *( ++i ), Camera_getOrigin( camwnd ) ) ) );
 		Vector3 angles( Camera_getAngles( camwnd ) );
-		angles[CAMERA_YAW] = static_cast<float>( radians_to_degrees( atan2( dir[1], dir[0] ) ) );
-		angles[CAMERA_PITCH] = static_cast<float>( radians_to_degrees( asin( dir[2] ) ) );
+		angles[CAMERA_YAW] = radians_to_degrees( atan2( dir[1], dir[0] ) );
+		angles[CAMERA_PITCH] = radians_to_degrees( asin( dir[2] ) );
 		Camera_setAngles( camwnd, angles );
 	}
 }
